+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Other changes
 
 * Fixed an issue where the `NavigationViewControllerDelegate.navigationViewController(_:didArriveAt:)` was not called for intermediate waypoints.
+* Fixed an issue where the `distanceUnit` passed to the `NavigationRouteOptions` and `NavigationMatchOptions` initializers was not correctly applied to the `distanceMeasurementSystem`.
 
 ## 3.10.0-beta.1